Crate cuid

cuid-rust

CUID generation in rust

use cuid;

// Get a full CUID
println!("{}", cuid::cuid().unwrap());

// Get a shorter, URL-appropriate CUID slug
println!("{}", cuid::slug().unwrap());

Note that this crate also provides a very simple, single purpose commandline interface:

$> cuid
ckfritrvg0000kdtwc766fful

See the original implementation for more details on CUIDs in general.

Enums

CuidError

Errors for the CUID library

Functions

cuid

Generate a CUID

is_cuid

Return whether a string is a legitimate CUID

is_slug

Return whether a string is a legitimate CUID slug

slug

Generate a CUID slug